requirements.txt
discord.py
python-dotenv
openai
pillow
customtkinter
packaging
